Face rendering + deleted broken backend cod
Wanted to get back to a runnable state after dinking around with a bunch
of stuff, so I reverted a lot of breaking/uninformed/stupid changes that
I made in the backend so I could focus on something easier: getting
faces to render on the frontend. Given the expected JSON representation
of a face (an example of which is in `fakeFaceJSON` in `index.html`)
this can render that face.